moving react dependencies into sub-folder
