+++ packages/headless/react.d.ts
@@ -1,0 +1,1 @@
+export * from './dist/react';
