+++ packages/headless/src/index.ts
@@ -6,8 +6,4 @@ export * from './auth';
 export * from './config';
 // ESLint thinks there's a conflict with ./types below due to Menu and MenuItem
 // on WPGraphQL
-// eslint-disable-next-line import/export
-export * from './components';
-export * from './provider';
-// eslint-disable-next-line import/export
 export * from './types';
