+++ packages/headless/src/react/provider/HeadlessProvider.tsx
@@ -1,7 +1,7 @@
 import React, { useMemo } from 'react';
 import { ApolloProvider, DocumentNode, gql } from '@apollo/client';
 import { useApollo } from './apolloClient';
-import type { ContentNodeOptions, ListPostOptions } from '../api';
+import type { ContentNodeOptions, ListPostOptions } from '../../api';
 
 export interface QueriesConfig {
   posts?: ListPostOptions;
