+++ packages/headless/src/react/provider/index.ts
@@ -1,3 +1,2 @@
 export * from './apolloClient';
 export * from './HeadlessProvider';
-export { gql, useQuery, useMutation } from '@apollo/client';
